data_ingestion/requirements.txt
yfinance
pandas
APScheduler
pyyaml
